Simplified ResetDialogPreference. #241
