+++ collect_app/src/main/java/org/odk/collect/android/preferences/ResetDialogPreference.java
@@ -1,37 +1,23 @@
 package org.odk.collect.android.preferences;
 
 import android.app.Activity;
-import android.app.AlarmManager;
 import android.app.AlertDialog;
-import android.app.PendingIntent;
 import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.preference.DialogPreference;
-import android.preference.PreferenceManager;
 import android.util.AttributeSet;
-import android.util.Log;
 import android.view.View;
 import android.widget.CheckBox;
 import android.widget.Toast;
 
-import org.jdeferred.Deferred;
-import org.jdeferred.DoneCallback;
-import org.jdeferred.DonePipe;
-import org.jdeferred.FailCallback;
-import org.jdeferred.Promise;
-import org.jdeferred.impl.DeferredObject;
 import org.odk.collect.android.R;
-import org.odk.collect.android.listeners.DeleteInstancesListener;
-import org.odk.collect.android.provider.DatabaseReader;
-import org.odk.collect.android.tasks.DeleteInstancesTask;
+import org.odk.collect.android.utilities.ResetResultCallback;
+import org.odk.collect.android.utilities.ResetUtility;
 
 public class ResetDialogPreference extends DialogPreference implements
-        DialogInterface.OnClickListener {
+        DialogInterface.OnClickListener, ResetResultCallback {
     final static String TAG = "ResetDialogPreference";
-
 
     CheckBox mPreferences;
     CheckBox mInstances;
@@ -45,10 +31,31 @@ public class ResetDialogPreference extends DialogPreference implements
     }
 
     @Override
+    public void onBindDialogView(View view) {
+        mPreferences = (CheckBox) view.findViewById(R.id.preferences);
+        mInstances = (CheckBox) view.findViewById(R.id.instances);
+
+        super.onBindDialogView(view);
+    }
+
+    @Override
     public void onClick(DialogInterface dialog, int which) {
         if (which == DialogInterface.BUTTON_POSITIVE) {
             resetSelected();
         }
+    }
+
+    @Override
+    public void doneResetting() {
+        hideProgressDialog();
+        showSuccessMessage();
+        ((Activity) mContext).finish();
+    }
+
+    @Override
+    public void failedToReset(String errorMessage) {
+        hideProgressDialog();
+        showErrorMessage(errorMessage);
     }
 
     private void resetSelected() {
@@ -62,59 +69,16 @@ public class ResetDialogPreference extends DialogPreference implements
 
         showProgressDialog();
 
-        Deferred deferred = new DeferredObject<Void, String, Void>();
-        Promise promise = deferred.promise();
-
-        if (resetPreferences) {
-            promise = promise
-                    .then(new DonePipe() {
-                        @Override
-                        public Deferred pipeDone(Object result) {
-                            resetPreferences();
-                            return new DeferredObject().resolve(null);
-                        }
-                    });
-        }
-
-        if (resetInstances) {
-            promise = promise.then(new DonePipe() {
-                @Override
-                public Promise pipeDone(Object result) {
-                    DeferredObject dobj = new DeferredObject<>();
-                    resetInstances(dobj);
-
-                    return dobj;
-                }
-            });
-        }
-
-        promise
-                .done(new DoneCallback() {
-                    @Override
-                    public void onDone(Object result) {
-                        hideProgressDialog();
-                        showSuccessMessage();
-                        ((Activity) mContext).finish();
-                    }
-                })
-                .fail(new FailCallback<String>() {
-                    @Override
-                    public void onFail(String errorMessage) {
-                        hideProgressDialog();
-                        showErrorMessage(errorMessage);
-                    }
-                });
-
-        deferred.resolve(null);
-    }
-
-    private void hideProgressDialog() {
-        mProgressDialog.dismiss();
+        new ResetUtility().reset(getContext(), resetPreferences, resetInstances, this);
     }
 
     private void showProgressDialog() {
         mProgressDialog = ProgressDialog.show(getContext(), "Please wait ...", "Resetting...",
                 true);
+    }
+
+    private void hideProgressDialog() {
+        mProgressDialog.dismiss();
     }
 
     private void showSuccessMessage() {
@@ -128,41 +92,5 @@ public class ResetDialogPreference extends DialogPreference implements
                 .setMessage(errorMessage)
                 .setPositiveButton(R.string.ok, null)
                 .show();
-    }
-
-    private void resetInstances(final DeferredObject deferred) {
-        final Long[] allInstances = new DatabaseReader().getAllInstancesIds(getContext());
-
-        DeleteInstancesTask task = new DeleteInstancesTask();
-        task.setContentResolver(getContext().getContentResolver());
-        task.setDeleteListener(new DeleteInstancesListener() {
-            @Override
-            public void deleteComplete(int deletedInstances) {
-                if (deletedInstances == allInstances.length) {
-                    deferred.resolve(null);
-                } else {
-                    deferred.reject(
-                            String.format("We've been able to delete only %d instances out of %d",
-                                    deletedInstances, allInstances.length));
-                }
-            }
-        });
-
-        task.execute(allInstances);
-    }
-
-    private void resetPreferences() {
-        PreferenceManager.getDefaultSharedPreferences(getContext())
-                .edit()
-                .clear()
-                .commit();
-    }
-
-    @Override
-    public void onBindDialogView(View view) {
-        mPreferences = (CheckBox) view.findViewById(R.id.preferences);
-        mInstances = (CheckBox) view.findViewById(R.id.instances);
-
-        super.onBindDialogView(view);
     }
 }
