+++ collect_app/src/main/java/org/odk/collect/android/utilities/ResetResultCallback.java
@@ -1,0 +1,6 @@
+package org.odk.collect.android.utilities;
+
+public interface ResetResultCallback {
+    void doneResetting();
+    void failedToReset(String errorMessage);
+}
